width auto
On line 195, delete the property "width: auto".

That property causes trouble when using "grid__col-s-". For example, when using "grid__col grid__col-6-of-12 grid__col-s-6-of-12", the 6 column on small device don't take the half of the grid! That problem is caused by "width: auto" in "@media (max-width: $toast-breakpoint-medium)" (line 191).